download command also loads graph into variable
